+++ PremiumScraps/Utils/ShipInventoryConditions.cs
@@ -14,13 +14,16 @@ namespace PremiumScraps.Utils
             }
         }
 
-        public static bool PremiumScrapsCondition(PlayerControllerB player)
+        private static bool PremiumScrapsCondition(PlayerControllerB player)
         {
             var item = player.currentlyHeldObjectServer;
-            if ((item.itemProperties.name == "BombItem" && item is Bomb bomb && bomb.activated) ||
-                (item.itemProperties.name == "ControllerItem" && item is Controller controller && controller.isInControlMode) ||
+            if ((item.itemProperties.name == "BombItem" && item is Bomb) ||
+                (item.itemProperties.name == "ControllerItem" && item is Controller) ||
                 (item.itemProperties.name == "JobApplicationItem" && item is JobDark) ||
-                (item.itemProperties.name == "GazpachoItem" && item is SpanishDrink))
+                (item.itemProperties.name == "GazpachoItem" && item is SpanishDrink) ||
+                (item.itemProperties.name == "ScrollItem" && item is ScrollTP) ||
+                (item.itemProperties.name == "AbiItem" && item is TalkingBall) ||
+                (item.itemProperties.name == "CustomFaceItem" && item is TrollFace))
                 return false;
             return true;
         }